requirements.txt
asgiref==3.8.1
boltons==23.0.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==39.0.1
Django==5.0.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.4
jsonpatch==1.32
jsonpointer==2.1
mod-wsgi==5.0.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycosat
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
regex==2023.6.3
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.5.0
tiktoken==0.4.0
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.11.0
urllib3==1.26.15
zstandard==0.19.0
